Title Screen Update

+++ src/mae/game/object/Obj_Skeleton_Card_Gen.java
@@ -3,6 +3,7 @@ package mae.game.object;
 import java.awt.Graphics2D;
 import java.awt.Rectangle;
 import java.awt.image.BufferedImage;
+import java.text.DecimalFormat;
 
 import mae.game.EntityType;
 import mae.game.GamePanel;
@@ -27,12 +28,12 @@ public class Obj_Skeleton_Card_Gen extends Object {
 	public void interact() {
 		//Remove 2 cards + Card should not be skeleton
 		if(slot1 == -1 && gp.currentCard != 0) {
-			slot1 = gp.kc[gp.currentCard].useCard();
+			slot1 = (int) gp.kc[gp.currentCard].useCard();
 			gp.kc[gp.currentCard].opened = false;
 			gp.currentCard = gp.kc[gp.currentCard].nextCard(gp.currentCard);
 		}
 		else if(slot2 == -1 && gp.currentCard != 0) {
-			slot2 = gp.kc[gp.currentCard].useCard();
+			slot2 = (int) gp.kc[gp.currentCard].useCard();
 			gp.kc[gp.currentCard].opened = false;
 			gp.currentCard = gp.kc[gp.currentCard].nextCard(gp.currentCard);
 		}
@@ -42,7 +43,7 @@ public class Obj_Skeleton_Card_Gen extends Object {
 			case PLUS: gp.kc[0].setValue(slot1 + slot2); break;
 			case MINUS: gp.kc[0].setValue(slot1 - slot2); break;
 			case MULTI: gp.kc[0].setValue(slot1 * slot2); break;
-			case DIVIDE: gp.kc[0].setValue(slot1 / slot2); break;
+			case DIVIDE: gp.kc[0].setValue((double)slot1 / (double)slot2); break;
 			}
 			gp.kc[slot1].opened = true;
 			slot1 = -1;
